Initial 3.0 changes
